[API change] homography returns a Matrix3d now
Before we used an Affine2d to represent a homography.
While this suits the usual use case, Eigen expects
that the last row is 0, 0, 1, which doesn't need to be case.
This might result in surprising bugs, e.g. when inverting a
homography.
On the other the homography is now more cumbersome to use.